fix: correção do princípio de SRP em get-transactions-statistics
