+++ src/repositories/transactions-repository.ts
@@ -1,13 +1,5 @@
 import type { Transaction } from '@/@types/transaction'
 
-interface GetTransactionsStatistics {
-  count: number
-  sum: number
-  avg: number
-  min: number
-  max: number
-}
-
 export interface TransactionsRepository {
   create(transaction: Transaction): Promise<void>
   deleteAll(): Promise<void>
@@ -15,5 +7,5 @@ export interface TransactionsRepository {
     rangeInMinutes,
   }: {
     rangeInMinutes?: number | null
-  }): Promise<GetTransactionsStatistics>
+  }): Promise<Transaction[]>
 }
